fix # Reform with Kotlin, step 20. Signed-off-by: wu.shengbo <wu.shb@reachauto.com>

+++ app/src/main/java/com/warchaser/musicplayer/tools/MusicList.kt
@@ -125,7 +125,10 @@ class MusicList private constructor(){
         }
 
         @JvmStatic
-        fun deleteSingleMusicFile(uri : String) : Boolean{
+        fun deleteSingleMusicFile(uri : String?) : Boolean{
+            if(uri == null){
+                return false
+            }
             return File(uri).delete()
         }
